requirements.txt
unstructured
tiktoken
pinecone
pypdf
groq
langchain
pandas
numpy
python-dotenv
streamlit
requests